Move manage specimens JSP and choose importer JSP/action
