+++ api/src/org/labkey/api/study/SpecimenUrls.java
@@ -39,4 +39,6 @@ public interface SpecimenUrls extends UrlProvider
     ActionURL getUpdateSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
 
     Class<? extends Controller> getCopyParticipantCommentActionClass();
+    Class<? extends Controller> getManageLocationTypesActionClass();
+    Class<? extends Controller> getManageSpecimenCommentsActionClass();
 }
