+++ specimen/src/org/labkey/specimen/actions/SpecimenController2.java
@@ -68,7 +68,6 @@ import org.labkey.api.specimen.model.ExtendedSpecimenRequestView;
 import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
-import org.labkey.specimen.notifications.ActorNotificationRecipientSet;
 import org.labkey.api.specimen.notifications.DefaultRequestNotification;
 import org.labkey.api.specimen.notifications.NotificationRecipientSet;
 import org.labkey.api.specimen.query.SpecimenQueryView;
@@ -94,6 +93,7 @@ import org.labkey.api.specimen.view.SpecimenWebPart;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.MapArrayExcelWriter;
 import org.labkey.api.study.SpecimenService;
+import org.labkey.api.study.SpecimenTransform;
 import org.labkey.api.study.SpecimenUrls;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyInternalService;
@@ -127,6 +127,7 @@ import org.labkey.api.view.VBox;
 import org.labkey.api.view.ViewBackgroundInfo;
 import org.labkey.api.view.ViewContext;
 import org.labkey.api.view.WebPartView;
+import org.labkey.specimen.notifications.ActorNotificationRecipientSet;
 import org.labkey.specimen.pipeline.SpecimenArchive;
 import org.labkey.specimen.pipeline.SpecimenBatch;
 import org.labkey.specimen.query.SpecimenEventQueryView;
@@ -4991,6 +4992,63 @@ public class SpecimenController2 extends SpringActionController
         public ActionURL getSuccessURL(EmailSpecimenListForm emailSpecimenListForm)
         {
             return getManageRequestURL(emailSpecimenListForm.getId(), null);
+        }
+    }
+
+    public static class EnabledSpecimenImportForm
+    {
+        private String _activeTransform;
+
+        public String getActiveTransform()
+        {
+            return _activeTransform;
+        }
+
+        public void setActiveTransform(String activeTransform)
+        {
+            _activeTransform = activeTransform;
+        }
+    }
+
+    @RequiresPermission(AdminPermission.class)
+    public class ChooseImporterAction extends FormViewAction<EnabledSpecimenImportForm>
+    {
+        @Override
+        public void addNavTrail(NavTree root)
+        {
+            root.addChild("Specimen Import Mechanism");
+        }
+
+        @Override
+        public void validateCommand(EnabledSpecimenImportForm target, Errors errors)
+        {
+        }
+
+        @Override
+        public ModelAndView getView(EnabledSpecimenImportForm form, boolean reshow, BindException errors) throws Exception
+        {
+            setHelpTopic(new HelpTopic("externalSpecimens"));
+            return new JspView<>("/org/labkey/specimen/view/chooseImporter.jsp", form, errors);
+        }
+
+        @Override
+        public boolean handlePost(EnabledSpecimenImportForm form, BindException errors) throws Exception
+        {
+            PropertyManager.PropertyMap props = PropertyManager.getWritableProperties(getContainer(), "enabledSpecimenImporter", true);
+            props.put("active", form.getActiveTransform());
+            props.save();
+            return true;
+        }
+
+        @Override
+        public URLHelper getSuccessURL(EnabledSpecimenImportForm configForm)
+        {
+            Container c = getContainer();
+            SpecimenService specimenService = SpecimenService.get();
+
+            String active = specimenService.getActiveSpecimenImporter(c);
+            SpecimenTransform activeTransform = specimenService.getSpecimenTransform(active);
+            return activeTransform.getManageAction(c, getUser());
         }
     }
 }
